requirements.txt
absl-py==0.7.1
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
aniso8601==7.0.0
anykeystore==0.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.8
autocorrect==0.3.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bert==2.2.0
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
blis==0.2.4
bokeh==1.0.4
boto==2.49.0
boto3==1.9.210
botocore==1.12.210
Bottleneck==1.2.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
conda==4.6.14
conda-build==3.17.8
conda-verify==3.1.1
contextlib2==0.5.5
coremltools==3.0b3
cryptacular==1.5.5
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.6
cytoolz==0.9.0.1
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
distributed==1.26.0
docutils==0.14
entrypoints==0.3
erlastic==2.0.0
et-xmlfile==1.0.1
fanstatic==1.1
fastai==1.0.57
fastcache==1.0.2
fastprogress==0.1.21
filelock==3.0.10
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
future==0.17.1
gast==0.2.2
gevent==1.4.0
glob2==0.6
gmpy2==2.0.8
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
hupper==1.8.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.0.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
js.d3==3.5.5
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
librosa==0.6.3
lief==0.9.0
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==3.0.5
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
murmurhash==1.0.2
mxnet==1.5.0
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
openpyxl==2.6.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.10
PasteDeploy==2.0.1
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pbkdf2==1.3
pdfminer.six==20181108
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pkginfo==1.5.0.1
plac==0.9.6
plaster==1.0
plaster-pastedeploy==0.7
plotly==3.10.0
pluggy==0.9.0
ply==3.11
preshed==2.0.1
prettytable==0.7.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.1
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
pycurl==7.43.0.2
pydub==0.23.1
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pynvx==1.0.0
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
PyPDF2==1.26.0
pyramid==1.10.4
pyramid-mailer==0.15.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytorch-pretrained-bert==0.6.2
-e git+https://github.com/eulertech/pytorch-transformers.git@40acf6b52a5250608c2b90edd955835131971d5a#egg=pytorch_transformers
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
regex==2019.6.8
repoze.sendmail==4.4.1
requests==2.21.0
requests-oauthlib==1.2.0
resampy==0.2.1
retrying==1.3.3
rope==0.12.0
ruamel-yaml==0.15.46
s3transfer==0.2.1
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
sentencepiece==0.1.83
shutilwhich==1.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
spacy==2.1.4
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
srsly==0.0.7
statsmodels==0.9.0
sympy==1.3
tables==3.5.1
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
thinc==7.0.4
tika==1.19
toolz==0.9.0
torch==1.2.0
torchvision==0.4.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
transaction==2.4.0
translationstring==1.3
turicreate==5.7
unicodecsv==0.14.1
urllib3==1.24.1
velruse==1.1.1
venusian==1.2.0
wasabi==0.2.2
Wave==0.0.2
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
WTForms==2.2.1
wtforms-recaptcha==0.3.2
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.4
xlwt==1.3.0
zict==0.1.4
zipp==0.3.3
zope.deprecation==4.4.0
zope.interface==4.6.0
zope.sqlalchemy==1.1
